BL2: Fix built in key id type and values
As per PSA Specification, the key ids must be of type `uint32_t` and
the values must be in the range of 'TFM_BUILTIN_KEY_ID_MIN' to
'TFM_BUILTIN_KEY_ID_MAX'.

Since the key IDs do not directly map 1:1 to the provisioned RoTPK
offsets in OTP, a lookup mechanism has been introduced to correctly
determine the OTP offset corresponding to each key ID.

This change also updates the logic for mapping PSA key IDs to the
corresponding key policy bit position in the BL2 ROTPK policy word.

Signed-off-by: Maulik Patel <maulik.patel@arm.com>
Change-Id: I63951d514776d0787dd3f4128edec59494e244ea