RSE: Refactor provisioning bin generation
Given that there are now 5 separate blob files to generate depending on
the configuration, with significant crossover in their CMake logic, this
can be tidied up by adding multiple functions.

Add two new CMake functions which are used to generate the provisioning
bundle executable target and to generate the final binary respectively.
These functions are designed to be generic and can be used to more
easily add support for future blobs. Currently, the generate binary
function is only used in the DUMMY_PROVISIONING case, but it could
potentially be used in future to generate production blobs as well.

Change-Id: I8879094ef08bef66ab42a5de852fa7b8087db84d
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>