platform: nordic_nrf: Add memory service header in common folder
When TF-M is build with Zephyr it now requires to have the platform
user memory ranges header.

This copies the memory range header into the common folder because
Zephyr builds from the common folder.

This is a workaround since the file is copied which is not ideal.
I wanted to have a minimal commit that can be applied cleanly to
the TF-M 2.2 branch and to the main branch and my intention is to
have this header file only in the common folder and remove the
duplicate later.

Change-Id: I0590eafb4212fbb4c1dabbb0c1b7594673e6b19b
Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
(cherry picked from commit 9e3750992106e27b9501b9ddb39446effc3881c2)