Build: Make incompatible-pointer-types warning
On GCC14, incompatible-pointer-types has been upgraded from warning to
an error, which causes build failures, particularly in the tests. Add a
command line argument to the compiler to revert back to the GCC13
behaviour and treat it as a warning instead.

Note that imcompatible-pointer-types is a warning that should definitely
be fixed as can lead to unexpected behaviour, but for now this at least
lets us build with GCC14.

Change-Id: I8c985121a9cba094eb923ca043066b64ee3bc35a
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>